Changes for app.codacy
Add curly braces around the nested statement(s) in this 'if' block.
if(!isValidating)
Remove this initialization to 'isValidating', the compiler will do that for you.
private bool isValidating = false;
Add curly braces around the nested statement(s) in this 'if' block.
if (isValidating)
Add curly braces around the nested statement(s) in this 'else' block.
else